+++ CHANGELOG.md
@@ -11,5 +11,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+## [1.1.0] - 2016-07-25
+### Added
+Added four new rules from ESLint v2.12.0.
+
+* [max-lines](http://eslint.org/docs/rules/max-lines) - Adds `maximumNumberOfLines` compatibility from JSCS.
+* [no-mixed-operators](http://eslint.org/docs/rules/no-mixed-operators) - Also required a change to `no-extra-parens`. Please see change section.
+* [object-curly-newline](http://eslint.org/docs/rules/object-curly-newline) - Adds `requirePaddingNewLinesInObjects` and `disallowPaddingNewLinesInObjects` compatibility from JSCS.
+* [rest-spread-spacing](http://eslint.org/docs/rules/rest-spread-spacing)
+
+### Changed
+* [no-extra-parens](http://eslint.org/docs/rules/no-extra-parens) - Turned the `conditionalAssign` option off.
+
 ## [1.0.0] - 2016-07-17
 - First release